Move LoggingDestination to :infrastructure:logging and remove logging's dependency on util.
