+++ infrastructure/logging/build.gradle
@@ -1,7 +1,5 @@
 dependencies {
     api 'org.apache.tuweni:tuweni-bytes'
 
-    implementation project(':util')
-
     implementation 'org.apache.logging.log4j:log4j-core'
 }
